requirements.txt
aioflask==0.4.0
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dataclasses-json==0.5.9
Flask==2.3.2
frozenlist==1.3.3
google-api-core==2.11.1
google-auth==2.21.0
google-cloud-appengine-logging==1.3.1
google-cloud-audit-log==0.2.5
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-logging==3.5.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
greenletio==0.10.1
grpc-google-iam-v1==0.12.6
grpcio==1.56.0
grpcio-status==1.56.0
h11==0.14.0
idna==3.4
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.210
langchainplus-sdk==0.0.19
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.4
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
proto-plus==1.22.3
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.10
PyYAML==6.0
requests==2.31.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.17
stripe==5.4.0
tenacity==8.2.2
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.22.0
Werkzeug==2.3.6
yarl==1.9.2
zipp==3.15.0
